evoClust_calc_rmsd.py: fix order of input files
OK, this `for` is very non-optimal.
Before taking for f in files outside this loop, it was how it should be.
"for f in files" was moved up because, in this way, the proper order of files is fixed.  Before the files in here were processed according to the mapping file.  Right now they will be processed as they are inputted to the program, so c1_ , c2, c3 etc.